Changed the hierarchy of files:
We do not compile the frontend typescript into javascript anymore, vite
take care of it.
The frontend fiels are used by vite in dev profile.
In t he prod profile, the vite cli on our machine compile the frontend
directory into a website directory in the /dockerfiles/nginx.
Nginx doesn't use a volume anymore, it just copy the content of the
website folder.
We have a vite config at our root, we can call npx vite build at the
root to build the website folder for nginx
Fixed the Makefile to take into account all the changes