Multiples adding
The ball rebounds are now calculating by the objects it colliders : the
Platform and Padding scripts. We do not use ShapeCast anymore but
trigger colliders.
The GameManager keeps track of the score and reset the ball we someone
score.
The Paddles now have a maximum range of movement. The value of that
range is given by the GameManager getPaddleMovementRange.
It allows us to have a range calculated depending on the number of
paddles.