Some chanegs for the ball speed
The bakk as an initial speed.
The paddels has a minimum and maximum speeds. The speeds is clamped when
hittign the paddle.
The ball initial speed (before hittign a paddle) is slower than the
paddle minimum speed. It allows the round to start slowly in case the
player is at the bottom.
The Ball now has a public reset methode, that the GameManager call.