The InputManager and the GameManager are now shared between client and server
The InputManager now has a list of PlyaerInputs, and external components
can setKeyDown or up. That way we can set those values from a message
from a client, from the server, from an AI, from an inputKey, with the
same InputManager.
Added the new class LocalInputs to manage the inputKeys.
The GameManager doesn't change the score now, it send a message in a
local message bus, and external scripts can receive it and do actions.
The ScoreTextUpdated catch the message and changes the score.