created the ServerPongGame class, it is the PongGame, but for server
Had to do a lot of tricks for it to work, from monkey patching to
sending invalid parameters.
The backend server is now a static server, serving the Server scenes
from the /scenes/ route.
Imported c3w-xmlhttprequest because babylonjs load functions uses it.
The havok .wasm is now dsirectlt read from the disk, with a direct path.
I will clean next commits.