Added Matchmaking, rooms and disconnection handling
The Matchmaking create a Room for 2 players at the bottom fo the waiting
queue.
The room is disposed when a player quit, and the other player is
notified.
The ServerPongGame is disposed when the Room is disposed, I will maybe
add a pool later.
The socket.data is now a class called SocketData.