Added the customComponent + Imported decorator + swc rollup
The imported decorator is only usable in a customScriptComponent child
class. It sets getter and setters to help the import from the unity
babylonjs toolkit.
It can take a parameter : a ZodType or a Constructor<T extends
CustomScriptComponent | ToolkitImported> with ToolkitImported being the
types that the unity babylonjs toolkit automatically casts like Material
or TransformNode.
If it s a ZodType, it checks that it can be parsed, if it is a
CustomScriptComponent, it try to retreive it from the transform with
GetComponent, if it is a ToolkitImported, it justs checks the type.
Each imported properties should have the decorator.
This commit add types safety to the dependency injection.